More cleanup of base.hpp
